Move Material description component into Material folder
It was in its own separate folder, but all the undercomponents of
material page should be in the Material folder.
