+++ src/components/material/MaterialDescription.tsx
